Using MIDI terminology.
The MIDI spec, as far as I know, seems to call them banks and programs,
not presets nor patches.  Let's go with that to avoid confusion.

I see now, MIDI has a 14-bit number (two 7-bit numbers) for the bank,
and on top of that, a 7-bit number for the program.  The DrumStation
should more than likely ignore the bank switching command entirely.