Add QuadraVerb Plus ReaBank
This is the second of these which I've added (the first being the
DrumStation) for which I actually own the device in question.  Alas,
the LCD screen is missing some columns of pixels, and the previous
owner might have overwritten some presets.  This is my best attempt
right now to note what the presets are *probably* named.

Edit: I've now restored the factory default presets, and after a
cursory glance, can't see any difference.  It looks like the previous
owner/s didn't override any of the presets, or restored them all
before shipping the unit to me.

This listing is therefore very likely accurate.